backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appdirs==1.4.4
attrs==26.1.0
bcrypt==5.0.0
blinker==1.9.0
build==1.4.3
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.7
click==8.3.2
cryptography==46.0.7
dashscope==1.25.17
dataclasses-json==0.6.7
datasets==4.8.4
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.18.0
durationpy==0.10
fastapi==0.135.3
filelock==3.28.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.10.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==7.1.0
instructor==1.15.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==1.2.15
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.30
langchain-openai==1.1.13
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.32
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
mypy_extensions==1.1.0
narwhals==2.19.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.4
oauthlib==3.3.1
onnxruntime==1.24.4
openai==2.32.0
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
orjson==3.11.8
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.1
pandas==3.0.2
pgvector==0.4.2
pillow==12.2.0
propcache==0.4.1
protobuf==6.33.6
psycopg2-binary==2.9.11
pyarrow==23.0.1
pybase64==1.4.3
pycparser==3.0
pydantic==2.13.1
pydantic-settings==2.13.1
pydantic_core==2.46.1
pydeck==0.9.1
Pygments==2.20.0
pypdf==6.10.2
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
ragas==0.4.3
rank-bm25==0.2.2
redis==7.4.0
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.4
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scikit-network==0.33.5
scipy==1.17.1
sentence-transformers==5.4.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
streamlit==1.56.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.1
zstandard==0.25.0
